I organized all the packages and created your daily questions
