+++ src/Coding_Bad_day_Questions/String2/PlusOut.java
@@ -1,4 +1,4 @@
-package Coding_Bad_day_Questions;
+package Coding_Bad_day_Questions.String2;
 
 public class PlusOut {
 //  Given a string and a non-empty word string, return a version of the original String where all chars have been replaced by pluses ("+"), except for appearances of the word string which are preserved unchanged.
