+++ src/Coding_Bad_day_Questions/alper_question_08_27/MakeBricks.java
@@ -1,0 +1,12 @@
+package Coding_Bad_day_Questions.alper_question_08_27;
+
+public class MakeBricks {
+    /*
+    We want to make a row of bricks that is goal inches long. We have a number of small bricks (1 inch each) and big bricks (5 inches each). Return true if it is possible to make the goal by choosing from the given bricks. This is a little harder than it looks and can be done without any loops. See also: Introduction to MakeBricks
+
+
+makeBricks(3, 1, 8) → true
+makeBricks(3, 1, 9) → false
+makeBricks(3, 2, 10) → true
+     */
+}
