+++ src/Coding_Bad_day_Questions/alper_question_08_28/LoneSum.java
@@ -1,0 +1,12 @@
+package Coding_Bad_day_Questions.alper_question_08_28;
+
+public class LoneSum {
+    /*
+Given 3 int values, a b c, return their sum. However, if one of the values is the same as another of the values, it does not count towards the sum.
+
+
+loneSum(1, 2, 3) → 6
+loneSum(3, 2, 3) → 2
+loneSum(3, 3, 3) → 0
+     */
+}
