+++ src/Coding_Bad_day_Questions/alper_question_08_29/DoubleChar.java
@@ -1,0 +1,12 @@
+package Coding_Bad_day_Questions.alper_question_08_29;
+
+public class DoubleChar {
+    /*
+    Given a string, return a string where for every char in the original, there are two chars.
+
+
+doubleChar("The") → "TThhee"
+doubleChar("AAbb") → "AAAAbbbb"
+doubleChar("Hi-There") → "HHii--TThheerree"
+     */
+}
