+++ src/Coding_Bad_day_Questions/alper_question_08_30/CountHi.java
@@ -1,0 +1,12 @@
+package Coding_Bad_day_Questions.alper_question_08_30;
+
+public class CountHi {
+   /* Return the number of times that the string "hi" appears anywhere in the given string.
+
+
+    countHi("abc hi ho") → 1
+    countHi("ABChi hi") → 2
+    countHi("hihi") → 2
+
+    */
+}
