+++ src/Coding_Bad_day_Questions/emre_question_08_27/WordEnds.java
@@ -1,4 +1,4 @@
-package Coding_Bad_day_Questions.EmreQuestioan_08_27;
+package Coding_Bad_day_Questions.emre_question_08_27;
 
 public class WordEnds {
 //    Given a string and a non-empty word string, return a string made of each char just before and just after every appearance of the word in the string. Ignore cases where there is no char before or after the word, and a char may be included twice if it is between two words.
